[kernel: fix timer conversion]
issue:

Flaky time syscall tests, with the latter of two time() calls returning
a smaller time value than the first one.

reason:

Integer overflow in the kernel time conversion function from
microseconds to time_t.

When converting the time, the time_t milliseconds fraction is added to
the microseconds by multiplying it with 1000. Since the milliseconds are
stored as an 16 bit integer the result of the multiplication is also a
16 bit value. That is, all values larger than 64 will cause an overflow
and subsequently yield an overall wrong time_t value.

fix:

Cast the milliseconds fraction of time_t to a 32 bit integer before
multiplying it.

+++ kernel/timer/timer.c
@@ -97,7 +97,7 @@ static size_t to_ticks(uint32_t us){
 static void to_time(void){
 	mutex_lock(&timer_mtx);
 
-	time_us += time.ms * 1000 + time.us;
+	time_us += (uint32_t)time.ms * 1000 + time.us;
 
 	time.us = time_us % 1000;
 	time.ms = (time_us / 1000) % 1000;
